+++ src/repositories/Product.js
@@ -10,6 +10,69 @@ class ProductRepository {
         const products = JSON.parse(productsJson);
         return products
     };
-}
 
+    getProductById = async (productId) => {
+        const products = await this.getProducts();
+        const product = products.find((p) => p.id === productId);
+        return product
+    };
+
+    addProduct = async (newProduct) => {
+        const products = await this.getProducts();
+        const newId =
+            products.length > 0 ? products[products.length - 1].id + 1 : 1;
+        const productToAdd = {
+            id: newId,
+            title: newProduct.title,
+            description: newProduct.description,
+            code: newProduct.code,
+            price: newProduct.price,
+            status: newProduct.status !== undefined ? newProduct.status : true,
+            stock: newProduct.stock,
+            category: newProduct.category,
+            thumbnails: newProduct.thumbnails || [],
+        };
+        products.push(productToAdd);
+        await fs.promises.writeFile(
+            this.path,
+            JSON.stringify(products, null, 2),
+            "utf-8"
+        );
+        return productToAdd
+    };
+
+    updateProductById = async (productId, updatedFields) => {
+        const products = await this.getProducts();
+        const productIndex = products.findIndex((p) => p.id === productId);
+        if (productIndex == -1)
+            return
+
+        if (updatedFields.id) {
+            delete updatedFields.id;
+        }
+        products[productIndex] = {
+            ...products[productIndex],
+            ...updatedFields
+        };
+        await fs.promises.writeFile(
+            this.path,
+            JSON.stringify(products, null, 2),
+            "utf-8"
+        );
+        return products[productIndex]
+    }
+
+
+    deleteProductById = async (productId) => {
+        const products = await this.getProducts();
+        const newProducts = products.filter((p) => p.id !== productId);
+        await fs.promises.writeFile(
+            this.path,
+            JSON.stringify(newProducts, null, 2),
+            "utf-8"
+        );
+        return true
+    };
+
+}
 export default new ProductRepository();
